Add -Z 'commands' option
Thanks to Olivier for the patch
Closes sourceforge patch #2
https://sourceforge.net/p/vgrabbj/patches/2/

"I need to make a lot of operations before sending the picture to the
ftp server (rotation, incrustation, etc). So I add an option to run all
needed tools from the inside of vgrabbj, via a system() call.

From the man page

-Z 'commands'
Ask vgrabbj to launch an external program. May be used to pass the
picture through ImageMagick or others filters. The first '%s' instance
in the command line will be replaced with filename. If UseTmpOut is On,
'commands' will be run before the file is renamed."